requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
sentence-transformers
pydantic
google-generativeai